[landscape] Move landscape logs, add server support
Landscape log collection has been in the logs plugin, this
moves it under the landscape plugin.  Previously the entire
directory was collected.  Now we just collect all the logs
that haven't been rotated.  "All logs" support was added in
case you want the whole directory.

Support was added for the landscape-server config.
Logs are covered by the same collection we have already.

The license.txt required some care.  It's a gpg signed file
but we don't need the actual license: just that it's not
malformed (gpg -v) and what the license details are (head).

Signed-off-by: Bryan Quigley <bryan.quigley@canonical.com>
Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>