[docker] Expand plugin options and include version info
This patch causes 'docker version' output to be gathered as part of the default
set of commands. Further it adds the 'docker-engine' package to the package
list as that is the current name of the community release of Docker.

Additionally, it makes the following changes to the plugin options:

- 'logs' will now gather container logs for running containers. This was
previously provided by the 'all' option.
- 'inspect' will capture 'docker inspect' output for running containers.
- 'all' will now cause 'logs' or 'inspect' to capture output from stopped
containers as well. Note that by itself, 'all' will not cause additional output
to be captured by the plugin.
- 'gather_all' will enable all above options. I.E. it will capture 'logs' and
'inspect' output for all containers, including stopped ones.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>