[Plugin] Apply log_size limitation by default to add_copy_spec()
Previously, add_copy_spec() would only apply a sizelimit if it was
explicitly set in the call. This in turn meant that the "default" log
size limit set by --log-size was not honored unless the plugin actually
read the option value and set it as the sizelimit parameter.

This patch changes this behavior so that the value of --log-size is
always applied to any add_copy_spec() call if a different sizelimit is
not specified by the individual call of add_copy_spec() or if --all-logs
is not enabled.

This patch also updates any plugin that was doing something akin to the
following to now instead accept the implied sizelimit, E.G.:

	limit = self.get_option('log_size')
	self.add_copy_spec('/some/log/file', sizelimit=limit)

becomes
	self.add_copy_spec('/some/log/file')

Any plugin that was setting a sizelimit other than the --log-size option
value remains unchanged.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>